Refactored Manager /items/[uuid] tests
